Refactored eventtype code into separate package
