+++ pkg/eventtype/eventtypes.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package broker
+package eventtype
 
 import (
 	"context"
